+++ pkg/validator/container_test.go
@@ -21,7 +21,6 @@ import (
 	conf "github.com/fairwindsops/polaris/pkg/config"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 )
 
 var resourceConfMinimal = `---
@@ -48,69 +47,7 @@ exemptions:
     - foo
 `
 
-var resourceConfRanges = `
-checks:
-  memoryRequestsRange: error
-  memoryLimitsRange: warning
-customChecks:
-  memoryLimitsRange:
-    containers:
-      exclude:
-      - initContainer
-    successMessage: Memory limits are within the required range
-    failureMessage: Memory limits should be within the required range
-    category: Resources
-    target: Container
-    schema:
-      '$schema': http://json-schema.org/draft-07/schema
-      type: object
-      required:
-      - resources
-      properties:
-        resources:
-          type: object
-          required:
-          - limits
-          properties:
-            limits:
-              type: object
-              required:
-              - memory
-              properties:
-                memory:
-                  type: string
-                  resourceMinimum: 200M
-                  resourceMaximum: 6G
-  memoryRequestsRange:
-    successMessage: Memory requests are within the required range
-    failureMessage: Memory requests should be within the required range
-    category: Resources
-    target: Container
-    containers:
-      exclude:
-      - initContainer
-    schema:
-      '$schema': http://json-schema.org/draft-07/schema
-      type: object
-      required:
-      - resources
-      properties:
-        resources:
-          type: object
-          required:
-          - requests
-          properties:
-            requests:
-              required:
-              - memory
-              properties:
-                memory:
-                  type: string
-                  resourceMinimum: 200M
-                  resourceMaximum: 3G
-`
-
-func testValidateResources(t *testing.T, container *corev1.Container, resourceConf *string, controllerName string, expectedErrors []*ResultMessage, expectedWarnings []*ResultMessage, expectedSuccesses []*ResultMessage) {
+func testValidate(t *testing.T, container *corev1.Container, resourceConf *string, controllerName string, expectedErrors []*ResultMessage, expectedWarnings []*ResultMessage, expectedSuccesses []*ResultMessage) {
 	cv := ContainerValidation{
 		Container:          container,
 		ResourceValidation: &ResourceValidation{},
@@ -188,151 +125,7 @@ func TestValidateResourcesEmptyContainer(t *testing.T) {
 
 	expectedSuccesses := []*ResultMessage{}
 
-	testValidateResources(t, &container, &resourceConfMinimal, "foo", expectedErrors, expectedWarnings, expectedSuccesses)
-}
-
-func TestValidateResourcesPartiallyValid(t *testing.T) {
-	cpuRequest, err := resource.ParseQuantity("100m")
-	assert.NoError(t, err, "Error parsing quantity")
-
-	cpuLimit, err := resource.ParseQuantity("200m")
-	assert.NoError(t, err, "Error parsing quantity")
-
-	container := corev1.Container{
-		Name: "Empty",
-		Resources: corev1.ResourceRequirements{
-			Requests: corev1.ResourceList{
-				"cpu": cpuRequest,
-			},
-			Limits: corev1.ResourceList{
-				"cpu": cpuLimit,
-			},
-		},
-	}
-
-	expectedWarnings := []*ResultMessage{
-		{
-			ID:       "memoryLimitsRange",
-			Type:     "warning",
-			Message:  "Memory limits should be within the required range",
-			Category: "Resources",
-		},
-	}
-
-	expectedErrors := []*ResultMessage{
-		{
-			ID:       "memoryRequestsRange",
-			Type:     "error",
-			Message:  "Memory requests should be within the required range",
-			Category: "Resources",
-		},
-	}
-
-	expectedSuccesses := []*ResultMessage{}
-
-	testValidateResources(t, &container, &resourceConfRanges, "foo", expectedErrors, expectedWarnings, expectedSuccesses)
-}
-
-func TestValidateResourcesInit(t *testing.T) {
-	cvEmpty := ContainerValidation{
-		Container:          &corev1.Container{},
-		ResourceValidation: &ResourceValidation{},
-	}
-	cvInit := ContainerValidation{
-		Container:          &corev1.Container{},
-		ResourceValidation: &ResourceValidation{},
-		IsInitContainer:    true,
-	}
-
-	parsedConf, err := conf.Parse([]byte(resourceConfRanges))
-	assert.NoError(t, err, "Expected no error when parsing config")
-
-	err = applyContainerSchemaChecks(&parsedConf, "", conf.Deployments, &cvEmpty)
-	if err != nil {
-		panic(err)
-	}
-	assert.Len(t, cvEmpty.Errors, 1)
-	assert.Len(t, cvEmpty.Warnings, 1)
-
-	err = applyContainerSchemaChecks(&parsedConf, "", conf.Deployments, &cvInit)
-	if err != nil {
-		panic(err)
-	}
-	assert.Len(t, cvInit.Errors, 0)
-}
-
-func TestValidateResourcesFullyValid(t *testing.T) {
-	cpuRequest, err := resource.ParseQuantity("300m")
-	assert.NoError(t, err, "Error parsing quantity")
-
-	cpuLimit, err := resource.ParseQuantity("400m")
-	assert.NoError(t, err, "Error parsing quantity")
-
-	memoryRequest, err := resource.ParseQuantity("400Mi")
-	assert.NoError(t, err, "Error parsing quantity")
-
-	memoryLimit, err := resource.ParseQuantity("500Mi")
-	assert.NoError(t, err, "Error parsing quantity")
-
-	container := corev1.Container{
-		Name: "Empty",
-		Resources: corev1.ResourceRequirements{
-			Requests: corev1.ResourceList{
-				"cpu":    cpuRequest,
-				"memory": memoryRequest,
-			},
-			Limits: corev1.ResourceList{
-				"cpu":    cpuLimit,
-				"memory": memoryLimit,
-			},
-		},
-	}
-
-	expectedSuccesses := []*ResultMessage{
-		{
-			ID:       "memoryRequestsRange",
-			Type:     "success",
-			Message:  "Memory requests are within the required range",
-			Category: "Resources",
-		},
-		{
-			ID:       "memoryLimitsRange",
-			Type:     "success",
-			Message:  "Memory limits are within the required range",
-			Category: "Resources",
-		},
-	}
-
-	testValidateResources(t, &container, &resourceConfRanges, "foo", []*ResultMessage{}, []*ResultMessage{}, expectedSuccesses)
-
-	expectedSuccesses = []*ResultMessage{
-		{
-			ID:       "cpuRequestsMissing",
-			Type:     "success",
-			Message:  "CPU requests are set",
-			Category: "Resources",
-		},
-		{
-			ID:       "memoryRequestsMissing",
-			Type:     "success",
-			Message:  "Memory requests are set",
-			Category: "Resources",
-		},
-		{
-			ID:       "cpuLimitsMissing",
-			Type:     "success",
-			Message:  "CPU limits are set",
-			Category: "Resources",
-		},
-		{
-			ID:       "memoryLimitsMissing",
-			Type:     "success",
-			Message:  "Memory limits are set",
-			Category: "Resources",
-		},
-	}
-
-	testValidateResources(t, &container, &resourceConfMinimal, "foo", []*ResultMessage{}, []*ResultMessage{}, expectedSuccesses)
+	testValidate(t, &container, &resourceConfMinimal, "foo", expectedErrors, expectedWarnings, expectedSuccesses)
 }
 
 func TestValidateHealthChecks(t *testing.T) {
@@ -1281,7 +1074,7 @@ func TestValidateResourcesExemption(t *testing.T) {
 	expectedErrors := []*ResultMessage{}
 	expectedSuccesses := []*ResultMessage{}
 
-	testValidateResources(t, &container, &resourceConfExemptions, "foo", expectedErrors, expectedWarnings, expectedSuccesses)
+	testValidate(t, &container, &resourceConfExemptions, "foo", expectedErrors, expectedWarnings, expectedSuccesses)
 
 	expectedWarnings = []*ResultMessage{
 		{
@@ -1315,22 +1108,8 @@ func TestValidateResourcesExemption(t *testing.T) {
 
 	disallowExemptionsConf := resourceConfExemptions + "\ndisallowExemptions: true"
 
-	testValidateResources(t, &container, &disallowExemptionsConf, "foo", expectedErrors, expectedWarnings, expectedSuccesses)
+	testValidate(t, &container, &disallowExemptionsConf, "foo", expectedErrors, expectedWarnings, expectedSuccesses)
 }
-
-/*
-func TestValidateResourceRangeExemption(t *testing.T) {
-	container := corev1.Container{
-		Name: "Empty",
-	}
-
-	expectedWarnings := []*ResultMessage{}
-	expectedErrors := []*ResultMessage{}
-	expectedSuccesses := []*ResultMessage{}
-
-	testValidateResources(t, &container, &resourceConfRangeExemptions, "foo", expectedErrors, expectedWarnings, expectedSuccesses)
-}
-*/
 
 func resetCV(cv ContainerValidation) ContainerValidation {
 	cv.Errors = []*ResultMessage{}
